requirements/requirements.txt
# ===================
# Web Framework
# ===================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# ===================
# PowerPoint Parsing
# ===================
python-pptx>=0.6.23
pdfminer.six>=20231228

# ===================
# LLM Integration
# ===================
langchain-core>=0.1.0
langchain-ollama>=0.0.2
ollama>=0.1.0

# ===================
# Text-to-Speech
# ===================
edge-tts>=6.1.9

# ===================
# Image Processing
# ===================
Pillow>=10.2.0

# ===================
# Language Detection
# ===================
langdetect>=1.0.9

# ===================
# Configuration
# ===================
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ===================
# Optional: Redis (for production caching)
# ===================
# redis>=5.0.0
# aioredis>=2.0.0
langchain-ollama
python-pptx
